# api/v1/openapi.yml
openapi: 3.2.0

# информация о документации
info:
  title: Wallet API
  description: API Specification for Wallet API
  version: 0.0.1

# теги для группировки эндпоинтов
tags:
  - name: wallet
    description: Manage wallets

# здесь описываются эндпоинты
paths:

  # путь для создания кошелька
  /api/v1/wallet/create:
    post:
      tags:
        - wallet
      summary: create wallet
      operationId: CreateWallet

      responses:
        '201':
          description: Wallet created
          headers:
            Location:
              description: URL of the created wallet
              schema:
                type: string
                format: uri
        '500':
          description: Internal error  # например, id по какой-то причине повторяется
          content:
            application/json:
              schema:
                # общий тип для ошибок
                $ref: "#/components/schemas/Error"
                example:
                  status: 500
                  title: Internal Server Error
                  detail: "Could not create wallet due to internal error."

  # внесение и снятие средств
  /api/v1/wallet:
    post:
      tags:
        - wallet
      summary: transfer money
      operationId: transfer

      requestBody:
        description: Money transfer information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'

      responses:
        '204':
          description: Transfer completed
        '400':
          description: Invalid input  # например, отрицательное число
          content:
            application/json:
              schema:
                # общий тип для ошибок
                $ref: "#/components/schemas/Error"
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # получение баланса
  /api/v1/wallets/{wallet_uuid}:
    get:
      tags:
        - wallet
      summary: get wallet balance
      operationId: getBalance

      parameters:
        - name: wallet_uuid
          in: path
          required: true
          description: UUID of wallet
          schema:
            type: string
            #format: uuid

      responses:
        '200':
          description: Got wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

# переиспользуемые объекты: схемы, типы ошибок, тела запросов и т.п.
components:
  schemas:

    Transfer:
      type: object
      properties:
        wallet_id:
          type: string
          #format: uuid
        operation:
          type: string
          enum: ["deposit", "withdraw"]
        amount:
          type: integer
          #format: int32
      required:
        - wallet_id
        - operation
        - amount

    Balance:
      type: object
      properties:
        wallet_id:
          type: string
          #format: uuid
        balance:
          type: integer
          #format: int32
      required:
        - wallet_id
        - balance

    Error:
      type: object
      properties:
        status:
          type: integer
          #format: int32
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
      required:
        - status
        - title
        - detail